Switch 'build' from phony to empty target.
Phony targets in make are really just targets that don't depend on
a file and are ALWAYS considered out-of-date. This means that even
if the target has been run once, it will run again, and again. This
breaks down for us when we try to build the collection of all Dockerfiles
as they track the dependencies appropriately but spend wasteful cycles
rebuilding the base images multiple times.

By switching to an empty target, we create a 'build' file in the
working directory where the Makefile is located and this 'build' file
now becomes a dependency on the 'build' target itself. In addition
to the 'build' file as a target dependency, we also depend on
Dockerfile and .env files, so the image will be rebuilt when these change.

In the future, the Makefile should support rebuilding when any arbitrary
file within the directory changes OR it's a force build.