Create 'build' and '$(BASE_IMAGE)-build' files to track changes.
A few things are being added in this commit.

1) The 'BUILD_DEPS' variable tracks all the files within the
directory of the Makefile. If any of the files are modified since
the last time the image was built, it will be marked stale and
re-created.

2) The 'IMAGEDEPS' variable renamed to 'BASE_IMAGE_DEPS'. This
still tracks the 'BASE_IMAGE' from .env to appropriate directory
and Makefile to build the image.

3) The automatic .PHONY targets for $(ALL), $(BUILD), and $(DEPLOY)
have been removed. These are being replaced with $(BASE_IMAGE)-%.
This target uses '%' so the 'all', 'build', or 'deploy' target can
be determined by the stem. This target now invokes the appropriate
Makefile in a child make process and creates a $(BASE_IMAGE)-stem
file in the working directory to indicate when the image was built.
This file is used to track _when_ the base image has been built and
when it needs to be rebuilt.

4) Move build-requirements into an order only dependency. This is a
logically .PHONY target (should be marked that way in the future)
as we want to run it on every invocation. .PHONY targets are always
out-of-date in terms of Make and if they're a normal target dependency,
which mark the target as out-of-date as well, causing _it_ to always
run every time. This is not what we want. By refactoring it to be
an order only dependency, it will not mark the build target as out-of-date.