Add 'make clean' target.
This is a recursive target and runs through the same DAG as the
other targets. Each Makefile should be responsible for cleaning up
any files created by previous build processes and any images in the
local docker cache.

TODO: Consider adding labels to images based on repo/image/tag
and including these in the filter when deleting.