Add arbitrary sleep before generating BUILD_TS.
Since our `date` command generates a timestamp from epoch in seconds,
when we use `touch` to set the *build files, we're setting it with
zero milliseconds/microseconds. Logically at the very _beginning_ of
the second. If there is not significant enough time between when
the `git checkout` finishes (setting the file last-modified timestamps)
and the `date` command, the build process will be touching files with a
date in the past, prior to the `git checkout` timeout and all builds will
be invalidated and rebuilt because the dependency files are newer than the
target file.