Refactor alpine images to new pattern that promotes DRY.
A number of things happened in this commit:

1) Common Dockerfile/Makefile code was moved into .common dir and
these files were symlinked into each versioned directory.

2) The common Makefile now specifies the Dockerfile to use in the
`docker build` command to the one in the common directory while using
the build context from the versioned directory.

3) Inherited environment variables from .env have moved from the parent
directory and into .common as well for consistency sake.

4) Makefile SUBDIR direction has been updated to exclude dot
directories in the scan. It was either this or renaming the common
ones .Dockerfile and .Makefile but felt the directory name was cleaner.