Refactor Image.make into Recursive.make.
Beyond the file rename, this generalizes the code that was identical
between Image.make and the Makefile in the root of the repository.

In general terms, this adds loading the root/parent .common dirs and
including the root common Makefile partials. With this, the differences
between the root Makefile and every other recursive makefile within
an image directory can be the same code.