Refactor $USER and $GROUP to $USR and $GROUP respectively.
If we set $USER as an envvar via the "ENV" directive in a Dockerfile,
it will be set prior to container launch by dockerd itself and not
from the actual shell process running within the container. This means
that if you run 'sudo su - root', $USER will still be whatever
dockerd set the value to and not 'root' as expected.

I don't know of any negative side-effects with $GROUP vs. $GRP at
this point but i'm refactoring the name for consistency sake.