Add release targets.
This commit adds the (patch|minor|major)-release targets to
Recursive.make that are supported only for invocated by the root
Makefile. These targets will bump the respective version number
in the VERSION file, commit the changes, create a git tag and
push these changes to the configured remote. Travis CI should be
configured to only perform its 'deploy' stage on tagged releases,
so this should trigger a 'make deploy' that pushes the built images
to the configured registry.