Support 'frozen' files.
If a image version directory contains a file named 'frozen', this
indicates that it should no longer be built using 'docker build'
as it's likely deprecated.

However, we still want the image to be part of this entire dockerfiles
build group. In this case, we pull down a copy of the image (latest tag),
tag it with values from this dockerfiles build execution and let those
tags get deployed.