Add 'skip' file support.
This enables directories to add a file named 'skip', in conjunction
with the 'frozen' file to skip the attempt at pulling down a remote
image as part of this build.

This functionality is necessary for forks of the repository that have
custom overrides that can no longer build older image versions
that are marked 'frozen'.

An alternative to this implementaiton would be to add .custom overrides
for these 'frozen' images in the fork to set the REPO value back to the
original repo's default value to pull them down. However, this will leave
the forked repo in an inconsistent state as not all images of the build pack
will be from the same base images.